requirements.txt
discord.py==1.3.4
matplotlib==3.3.4
requests
wordcloud
psycopg2
pytest
multidict
pyyaml
mee6-py-api
spacy==3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.0.0/en_core_web_md-3.0.0.tar.gz
